requirements.txt
amqp==5.3.1
aniso8601==10.0.0
async-timeout==5.0.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.1
lml==0.2.0
MarkupSafe==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.50
pyexcel==0.7.2
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.38
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
